refactor: 형변환 long 타입 변경

+++ src/main/java/blackjack/domain/card/Cards.java
@@ -25,8 +25,8 @@ public class Cards {
                 .sum();
     }
 
-    public int countAceCard() {
-        return (int) cards.stream()
+    public long countAceCard() {
+        return cards.stream()
                 .filter(Card::isAce)
                 .count();
     }
